+++ src/utils/extension.constants.ts
@@ -1,0 +1,7 @@
+import type { TExtensionConfig } from "../extension.types";
+
+export const CUSTOM_EXTENSION_CONFIG: TExtensionConfig = {
+  label: "Custom",
+  description: "Create multiple files based on input",
+  files: [],
+};